moved publish iteration logic to pipeline
