+++ client/ayon_core/tools/publisher/models/publish.py
@@ -1,557 +1,22 @@
 from __future__ import annotations
 
-import uuid
-import inspect
-import logging
 import traceback
-import collections
-from contextlib import contextmanager
 from functools import partial
-from typing import Any, Iterable, Generator
+from typing import Any, Iterable
 
 import pyblish.plugin
 
-from ayon_core.lib import env_value_to_bool
-from ayon_core.pipeline import (
-    PublishValidationError,
-    KnownPublishError,
-    OptionalPyblishPluginMixin,
+from ayon_core.pipeline.publish.logic import (
+    PublishIterInfo,
+    PublishLogic,
+    PublishActionResult,
+    PublishErrorInfo,
+    PublishErrorsReport,
 )
-from ayon_core.pipeline.publish import (
-    PublishError,
-    filter_crashed_publish_paths,
-)
-from ayon_core.pipeline.publish.report import (
-    PublishReportMaker,
-    PublishReport,
-)
+from ayon_core.pipeline.publish.report import PublishReport
 from ayon_core.tools.publisher.abstract import AbstractPublisherBackend
 
 PUBLISH_EVENT_SOURCE = "publisher.publish.model"
-# Define constant for plugin orders offset
-PLUGIN_ORDER_OFFSET = 0.5
-
-
-class MessageHandler(logging.Handler):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self._records = []
-
-    def clear_records(self):
-        self._records = []
-
-    def emit(self, record):
-        try:
-            record.msg = record.getMessage()
-        except Exception:
-            record.msg = str(record.msg)
-        self._records.append(record)
-
-    def get_records(self):
-        return self._records
-
-
-class PublishErrorInfo:
-    def __init__(
-        self,
-        message: str,
-        is_unknown_error: bool,
-        description: str | None = None,
-        title: str | None = None,
-        detail: str | None = None,
-    ):
-        self.message: str = message
-        self.is_unknown_error = is_unknown_error
-        self.description: str = description or message
-        self.title: str = title or "Unknown error"
-        self.detail: str | None = detail
-
-    def __eq__(self, other: Any) -> bool:
-        if not isinstance(other, PublishErrorInfo):
-            return False
-        return (
-            self.description == other.description
-            and self.is_unknown_error == other.is_unknown_error
-            and self.title == other.title
-            and self.detail == other.detail
-        )
-
-    def __ne__(self, other: Any) -> bool:
-        return not self.__eq__(other)
-
-    @classmethod
-    def from_exception(cls, exc) -> "PublishErrorInfo":
-        if isinstance(exc, PublishError):
-            return cls(
-                exc.message,
-                False,
-                exc.description,
-                title=exc.title,
-                detail=exc.detail,
-            )
-        if isinstance(exc, KnownPublishError):
-            msg = str(exc)
-        else:
-            msg = (
-                "Something went wrong. Send report"
-                " to your supervisor or Ynput team."
-            )
-        return cls(msg, True)
-
-
-class PublishPluginActionItem:
-    """Representation of publish plugin action.
-
-    Data driven object which is used as proxy for controller and UI.
-
-    Args:
-        action_id (str): Action id.
-        plugin_id (str): Plugin id.
-        active (bool): Action is active.
-        on_filter (Literal["all", "notProcessed", "processed", "failed",
-            "warning", "failedOrWarning", "succeeded"]): Actions have 'on'
-            attribute which define  when can be action triggered
-            (e.g. 'all', 'failed', ...).
-        label (str): Action's label.
-        icon (str | None) Action's icon.
-    """
-
-    def __init__(
-        self,
-        action_id: str,
-        plugin_id: str,
-        active: bool,
-        on_filter: str,
-        label: str,
-        icon: str | None,
-    ):
-        self.action_id: str = action_id
-        self.plugin_id: str = plugin_id
-        self.active: bool = active
-        self.on_filter: str = on_filter
-        self.label: str = label
-        self.icon: str | None = icon
-
-    def to_data(self) -> dict[str, str | bool | None]:
-        """Serialize object to dictionary.
-
-        Returns:
-            dict[str, str | bool | None]: Serialized object.
-
-        """
-        return {
-            "action_id": self.action_id,
-            "plugin_id": self.plugin_id,
-            "active": self.active,
-            "on_filter": self.on_filter,
-            "label": self.label,
-            "icon": self.icon
-        }
-
-    @classmethod
-    def from_data(
-        cls, data: dict[str, str | bool | None]
-    ) -> "PublishPluginActionItem":
-        """Create object from data.
-
-        Args:
-            data (dict[str, str | bool | None]): Data used to recreate
-                object.
-
-        Returns:
-            PublishPluginActionItem: Object created using data.
-        """
-
-        return cls(**data)
-
-
-class PublishPluginsProxy:
-    """Wrapper around publish plugin.
-
-    Prepare mapping for publish plugins and actions. Also can create
-    serializable data for plugin actions so UI don't have to have access to
-    them.
-
-    This object is created in process where publishing is actually running.
-
-    Notes:
-        Actions have id but single action can be used on multiple plugins so
-            to run an action is needed combination of plugin and action.
-
-    Args:
-        plugins [list[pyblish.api.Plugin]]: Discovered plugins that will be
-            processed.
-    """
-
-    def __init__(self, plugins: list[pyblish.api.Plugin]):
-        plugins_by_id: dict[str, pyblish.api.Plugin] = {}
-        actions_by_plugin_id: dict[str, dict[str, pyblish.api.Action]] = {}
-        action_ids_by_plugin_id: dict[str, list[str]] = {}
-        for plugin in plugins:
-            plugin_id = plugin.id
-            plugins_by_id[plugin_id] = plugin
-
-            action_ids = []
-            actions_by_id = {}
-            action_ids_by_plugin_id[plugin_id] = action_ids
-            actions_by_plugin_id[plugin_id] = actions_by_id
-
-            actions = getattr(plugin, "actions", None) or []
-            for action in actions:
-                action_id = action.id
-                action_ids.append(action_id)
-                actions_by_id[action_id] = action
-
-        self._plugins_by_id: dict[str, pyblish.api.Plugin] = plugins_by_id
-        self._actions_by_plugin_id: dict[
-            str, dict[str, pyblish.api.Action]
-        ] = actions_by_plugin_id
-        self._action_ids_by_plugin_id: dict[str, list[str]] = (
-            action_ids_by_plugin_id
-        )
-
-    def get_action(
-        self, plugin_id: str, action_id: str
-    ) -> pyblish.api.Action:
-        return self._actions_by_plugin_id[plugin_id][action_id]
-
-    def get_plugin(self, plugin_id: str) -> pyblish.api.Plugin:
-        return self._plugins_by_id[plugin_id]
-
-    def get_plugin_id(self, plugin: pyblish.api.Plugin) -> str:
-        """Get id of plugin based on plugin object.
-
-        It's used for validation errors report.
-
-        Args:
-            plugin (pyblish.api.Plugin): Publish plugin for which id should be
-                returned.
-
-        Returns:
-            str: Plugin id.
-        """
-
-        return plugin.id
-
-    def get_plugin_action_items(
-        self, plugin_id: str
-    ) -> list[PublishPluginActionItem]:
-        """Get plugin action items for plugin by its id.
-
-        Args:
-            plugin_id (str): Publish plugin id.
-
-        Returns:
-            list[PublishPluginActionItem]: Items with information about publish
-                plugin actions.
-        """
-
-        return [
-            self._create_action_item(
-                self.get_action(plugin_id, action_id), plugin_id
-            )
-            for action_id in self._action_ids_by_plugin_id[plugin_id]
-        ]
-
-    def _create_action_item(
-        self, action: pyblish.api.Action, plugin_id: str
-    ) -> PublishPluginActionItem:
-        label = action.label or action.__name__
-        icon = getattr(action, "icon", None)
-        return PublishPluginActionItem(
-            action.id,
-            plugin_id,
-            action.active,
-            action.on,
-            label,
-            icon
-        )
-
-
-class PublishErrorItem:
-    """Data driven publish error item.
-
-    Prepared data container with information about publish error and it's
-    source plugin.
-
-    Can be converted to raw data and recreated should be used for controller
-    and UI connection.
-
-    Args:
-        instance_id (str | None): Pyblish instance id to which is
-            publish error connected.
-        instance_label (str | None): Prepared instance label.
-        plugin_id (str): Pyblish plugin id which triggered the publish
-            error. Id is generated using 'PublishPluginsProxy'.
-        is_context_plugin (bool): Error happened on context.
-        title (str): Error title.
-        description (str): Error description.
-        detail (str): Error detail.
-
-    """
-    def __init__(
-        self,
-        instance_id: str | None,
-        instance_label: str | None,
-        plugin_id: str,
-        is_context_plugin: bool,
-        is_validation_error: bool,
-        title: str,
-        description: str,
-        detail: str
-    ):
-        self.instance_id: str | None = instance_id
-        self.instance_label: str | None = instance_label
-        self.plugin_id: str = plugin_id
-        self.is_context_plugin: bool = is_context_plugin
-        self.is_validation_error: bool = is_validation_error
-        self.title: str = title
-        self.description: str = description
-        self.detail: str = detail
-
-    def to_data(self) -> dict[str, str | bool | None]:
-        """Serialize object to dictionary.
-
-        Returns:
-            dict[str, str | bool | None]: Serialized object data.
-
-        """
-        return {
-            "instance_id": self.instance_id,
-            "instance_label": self.instance_label,
-            "plugin_id": self.plugin_id,
-            "is_context_plugin": self.is_context_plugin,
-            "is_validation_error": self.is_validation_error,
-            "title": self.title,
-            "description": self.description,
-            "detail": self.detail,
-        }
-
-    @classmethod
-    def from_result(
-        cls,
-        plugin_id: str,
-        error: PublishError,
-        instance: pyblish.api.Instance | None
-    ):
-        """Create new object based on resukt from controller.
-
-        Returns:
-            PublishErrorItem: New object with filled data.
-        """
-
-        instance_label = None
-        instance_id = None
-        if instance is not None:
-            instance_label = (
-                instance.data.get("label") or instance.data.get("name")
-            )
-            instance_id = instance.id
-
-        return cls(
-            instance_id,
-            instance_label,
-            plugin_id,
-            instance is None,
-            isinstance(error, PublishValidationError),
-            error.title,
-            error.description,
-            error.detail,
-        )
-
-    @classmethod
-    def from_data(cls, data):
-        return cls(**data)
-
-
-class PublishErrorsReport:
-    """Publish errors report that can be parsed to raw data.
-
-    Args:
-        error_items (list[PublishErrorItem]): List of publish errors.
-        plugin_action_items (dict[str, list[PublishPluginActionItem]]): Action
-            items by plugin id.
-
-    """
-    def __init__(self, error_items, plugin_action_items):
-        self._error_items = error_items
-        self._plugin_action_items = plugin_action_items
-
-    def __iter__(self) -> Iterable[PublishErrorItem]:
-        for item in self._error_items:
-            yield item
-
-    def group_items_by_title(self) -> list[dict[str, Any]]:
-        """Group errors by plugin and their titles.
-
-        Items are grouped by plugin and title -> same title from different
-        plugin is different item. Items are ordered by plugin order.
-
-        Returns:
-            list[dict[str, Any]]: List where each item title, instance
-                information related to title and possible plugin actions.
-        """
-
-        ordered_plugin_ids = []
-        error_items_by_plugin_id = collections.defaultdict(list)
-        for error_item in self._error_items:
-            plugin_id = error_item.plugin_id
-            if plugin_id not in ordered_plugin_ids:
-                ordered_plugin_ids.append(plugin_id)
-            error_items_by_plugin_id[plugin_id].append(error_item)
-
-        grouped_error_items = []
-        for plugin_id in ordered_plugin_ids:
-            plugin_action_items = self._plugin_action_items[plugin_id]
-            error_items = error_items_by_plugin_id[plugin_id]
-
-            titles = []
-            error_items_by_title = collections.defaultdict(list)
-            for error_item in error_items:
-                title = error_item.title
-                if title not in titles:
-                    titles.append(error_item.title)
-                error_items_by_title[title].append(error_item)
-
-            for title in titles:
-                grouped_error_items.append({
-                    "id": uuid.uuid4().hex,
-                    "plugin_id": plugin_id,
-                    "plugin_action_items": list(plugin_action_items),
-                    "error_items": error_items_by_title[title],
-                    "title": title
-                })
-        return grouped_error_items
-
-    def to_data(self):
-        """Serialize object to dictionary.
-
-        Returns:
-            dict[str, Any]: Serialized data.
-        """
-
-        error_items = [
-            item.to_data()
-            for item in self._error_items
-        ]
-
-        plugin_action_items = {
-            plugin_id: [
-                action_item.to_data()
-                for action_item in action_items
-            ]
-            for plugin_id, action_items in self._plugin_action_items.items()
-        }
-
-        return {
-            "error_items": error_items,
-            "plugin_action_items": plugin_action_items
-        }
-
-    @classmethod
-    def from_data(
-        cls, data: dict[str, Any]
-    ) -> "PublishErrorsReport":
-        """Recreate object from data.
-
-        Args:
-            data (dict[str, Any]): Data to recreate object. Can be created
-                using 'to_data' method.
-
-        Returns:
-            PublishErrorsReport: New object based on data.
-        """
-
-        error_items = [
-            PublishErrorItem.from_data(error_item)
-            for error_item in data["error_items"]
-        ]
-        plugin_action_items = {}
-        for action_item in data["plugin_action_items"]:
-            item = PublishPluginActionItem.from_data(action_item)
-            action_items = plugin_action_items.setdefault(item.plugin_id, [])
-            action_items.append(item)
-
-        return cls(error_items, plugin_action_items)
-
-
-class PublishErrors:
-    """Object to keep track about publish errors by plugin."""
-
-    def __init__(self):
-        self._plugins_proxy: PublishPluginsProxy | None = None
-        self._error_items: list[PublishErrorItem] = []
-        self._plugin_action_items: dict[
-            str, list[PublishPluginActionItem]
-        ] = {}
-
-    def __bool__(self):
-        return self.has_errors
-
-    @property
-    def has_errors(self) -> bool:
-        """At least one error was added."""
-
-        return bool(self._error_items)
-
-    def reset(self, plugins_proxy: PublishPluginsProxy):
-        """Reset object to default state.
-
-        Args:
-            plugins_proxy (PublishPluginsProxy): Proxy which store plugins,
-                actions by ids and create mapping of action ids by plugin ids.
-        """
-
-        self._plugins_proxy = plugins_proxy
-        self._error_items = []
-        self._plugin_action_items = {}
-
-    def create_report(self) -> PublishErrorsReport:
-        """Create report based on currently existing errors.
-
-        Returns:
-            PublishErrorsReport: Publish error report with all
-                error information and publish plugin action items.
-        """
-
-        return PublishErrorsReport(
-            self._error_items, self._plugin_action_items
-        )
-
-    def add_error(
-        self,
-        plugin: pyblish.api.Plugin,
-        error: PublishError,
-        instance: pyblish.api.Instance | None
-    ):
-        """Add error from pyblish result.
-
-        Args:
-            plugin (pyblish.api.Plugin): Plugin which triggered error.
-            error (PublishError): Publish error.
-            instance (pyblish.api.Instance | None): Instance on which was
-                error raised or None if was raised on context.
-        """
-
-        # Make sure the cached report is cleared
-        plugin_id = self._plugins_proxy.get_plugin_id(plugin)
-        if not error.title:
-            if hasattr(plugin, "label") and plugin.label:
-                plugin_label = plugin.label
-            else:
-                plugin_label = plugin.__name__
-            error.title = plugin_label
-
-        self._error_items.append(
-            PublishErrorItem.from_result(plugin_id, error, instance)
-        )
-        if plugin_id in self._plugin_action_items:
-            return
-
-        plugin_actions = self._plugins_proxy.get_plugin_action_items(
-            plugin_id
-        )
-        self._plugin_action_items[plugin_id] = plugin_actions
 
 
 def collect_families_from_instances(
@@ -589,227 +54,128 @@ class PublishModel:
     def __init__(self, controller: AbstractPublisherBackend):
         self._controller = controller
 
-        self._log_to_console: bool = env_value_to_bool(
-            "AYON_PUBLISHER_PRINT_LOGS", default=False
-        )
-
-        # Publishing should stop at validation stage
-        self._publish_up_validation: bool = False
-        self._publish_comment_is_set: bool = False
-
-        # Any other exception that happened during publishing
-        self._publish_error_info: PublishErrorInfo | None = None
-        # Publishing is in progress
-        self._publish_is_running: bool = False
-        # Publishing is over validation order
-        self._publish_has_validated: bool = False
-
-        self._publish_has_validation_errors: bool = False
-        self._publish_has_crashed: bool = False
-        # All publish plugins are processed
-        self._publish_has_started: bool = False
-        self._publish_has_finished: bool = False
-        self._publish_max_progress: int = 0
-        self._publish_progress: int = 0
-
-        self._publish_plugins: list[pyblish.api.Plugin] = []
-        self._publish_plugins_proxy: PublishPluginsProxy = (
-            PublishPluginsProxy([])
-        )
-
-        # pyblish.api.Context
-        self._publish_context = None
-        # Pyblish report
-        self._publish_report: PublishReportMaker = PublishReportMaker()
-        # Store exceptions of publish error
-        self._publish_errors: PublishErrors = PublishErrors()
-
-        # This information is not much important for controller but for widget
-        #   which can change (and set) the comment.
-        self._publish_comment_is_set: bool = False
-
-        # Validation order
-        # - plugin with order same or higher than this value is extractor or
-        #   higher
-        self._validation_order: int = (
-            pyblish.api.ValidatorOrder + PLUGIN_ORDER_OFFSET
-        )
-
-        # Plugin iterator
-        self._main_thread_iter: Generator[partial] = self._default_iterator()
-
-        self._log_handler: MessageHandler = MessageHandler()
+        self._logic: PublishLogic = PublishLogic()
 
     def reset(self):
-        # Allow to change behavior during process lifetime
-        self._log_to_console = env_value_to_bool(
-            "AYON_PUBLISHER_PRINT_LOGS", default=False
-        )
-
         create_context = self._controller.get_create_context()
 
-        self._publish_up_validation = False
-        self._publish_comment_is_set = False
-        self._publish_has_started = False
-
-        self._set_publish_error_info(None)
-        self._set_progress(0)
-        self._set_is_running(False)
-        self._set_has_validated(False)
-        self._set_is_crashed(False)
-        self._set_has_validation_errors(False)
-        self._set_finished(False)
-
-        self._main_thread_iter = self._publish_iterator()
-        self._publish_context = pyblish.api.Context()
-        # Make sure "comment" is set on publish context
-        self._publish_context.data["comment"] = ""
-        # Add access to create context during publishing
-        # - must not be used for changing CreatedInstances during publishing!
-        # QUESTION
-        # - pop the key after first collector using it would be safest option?
-        self._publish_context.data["create_context"] = create_context
-        publish_plugins = create_context.publish_plugins
-        self._publish_plugins = publish_plugins
-        self._publish_plugins_proxy = PublishPluginsProxy(
-            publish_plugins
-        )
-        blocking_crashed_paths = filter_crashed_publish_paths(
-            create_context.get_current_project_name(),
-            set(create_context.publish_discover_result.crashed_file_paths),
-            project_settings=create_context.get_current_project_settings(),
-        )
-        self._publish_report.reset_with_discover_results(
-            create_context.creator_discover_result,
-            create_context.convertor_discover_result,
-            create_context.publish_discover_result,
-            blocking_crashed_paths,
-        )
-        for plugin in create_context.publish_plugins_mismatch_targets:
-            self._publish_report.set_plugin_skipped(plugin.id)
-        self._publish_errors.reset(self._publish_plugins_proxy)
-
-        self._set_max_progress(len(publish_plugins))
+        self._logic.reset_from_create_context(create_context)
 
         self._emit_event("publish.reset.finished")
 
-    def set_publish_up_validation(self, value: bool):
-        self._publish_up_validation = value
+    def set_publish_up_validation(self, value: bool) -> None:
+        self._logic.set_publish_up_validation(value)
 
-    def start_publish(self, wait: bool = True):
+    def start_publish(self, wait: bool = True) -> None:
         """Run publishing.
 
         Make sure all changes are saved before method is called (Call
         'save_changes' and check output).
         """
-        if self._publish_up_validation and self._publish_has_validated:
+        if self._logic.is_running():
             return
 
-        self._start_publish()
+        self._emit_event("publish.process.started")
 
+        self._logic.start_publish(wait=False)
         if not wait:
             return
 
+        plugin = None
+        item_label = None
+        validated = False
         while self.is_running():
-            func = self.get_next_process_func()
+            func: PublishIterInfo = self.get_next_process_func()
+            if plugin is not func.plugin:
+                plugin = func.plugin
+                plugin_label = getattr(plugin, "label", None)
+                if not plugin_label:
+                    plugin_label = plugin.__name__
+                self._emit_event(
+                "publish.process.plugin.changed",
+                {"plugin_label": plugin_label}
+                )
+
+            if item_label != func.item_label:
+                item_label = func.item_label
+                self._emit_event(
+                    "publish.process.instance.changed",
+                    {"instance_label": item_label}
+                )
+
             func()
+            if not validated and self._logic.has_validated():
+                validated = True
+                self._emit_event("publish.has_validated")
 
-    def get_next_process_func(self) -> partial:
-        # Raise error if this function is called when publishing
-        #   is not running
-        if not self._publish_is_running:
-            raise ValueError("Publish is not running")
+        self.process_stopped()
 
-        # Validations of progress before using iterator
-        # Any unexpected error happened
-        # - everything should stop
-        if self._publish_has_crashed:
-            return partial(self.stop_publish)
+    def process_stopped(self):
+        self._emit_event("publish.process.stopped")
+        if self._logic.has_crashed():
+            pass
+        elif self._logic.has_finished():
+            self._emit_event("publish.finished")
 
-        # Stop if validation is over and validation errors happened
-        #   or publishing should stop at validation
-        if (
-            self._publish_has_validated
-            and (
-                self._publish_has_validation_errors
-                or self._publish_up_validation
-            )
-        ):
-            return partial(self.stop_publish)
+    def get_next_process_func(self) -> PublishIterInfo:
+        return self._logic.get_next_process_func()
 
-        # Everything is ok so try to get new processing item
-        return next(self._main_thread_iter)
-
-    def stop_publish(self):
-        if self._publish_is_running:
-            self._stop_publish()
+    def stop_publish(self) -> None:
+        self._logic.stop_publish()
 
     def is_running(self) -> bool:
-        return self._publish_is_running
+        return self._logic.is_running()
 
-    def is_crashed(self) -> bool:
-        return self._publish_has_crashed
+    def has_crashed(self) -> bool:
+        return self._logic.has_crashed()
 
     def has_started(self) -> bool:
-        return self._publish_has_started
+        return self._logic.has_started()
 
     def has_finished(self) -> bool:
-        return self._publish_has_finished
+        return self._logic.has_finished()
 
     def has_validated(self) -> bool:
-        return self._publish_has_validated
+        return self._logic.has_validated()
 
     def has_validation_errors(self) -> bool:
-        return self._publish_has_validation_errors
+        return self._logic.has_validation_errors()
 
     def publish_can_continue(self) -> bool:
-        return (
-            not self._publish_has_crashed
-            and not self._publish_has_validation_errors
-            and not self._publish_has_finished
-        )
+        return self._logic.publish_can_continue()
 
     def get_progress(self) -> int:
-        return self._publish_progress
+        return self._logic.get_progress()
 
     def get_max_progress(self) -> int:
-        return self._publish_max_progress
+        return self._logic.get_max_progress()
 
     def get_publish_report(self) -> PublishReport:
-        self._publish_report.update_publish_instances(self._publish_context)
-        return self._publish_report.get_report()
+        return self._logic.get_publish_report()
 
     def get_publish_report_data(self) -> dict[str, Any]:
-        report: PublishReport = self.get_publish_report()
-        return report.to_data(self._publish_report.current_plugin_id)
+        return self._logic.get_publish_report_data()
 
     def get_publish_errors_report(self) -> PublishErrorsReport:
-        return self._publish_errors.create_report()
+        return self._logic.get_publish_errors_report()
 
     def get_error_info(self) -> PublishErrorInfo | None:
-        return self._publish_error_info
+        return self._logic.get_error_info()
 
     def store_publish_report(self, filepath: str) -> None:
         report: PublishReport = self.get_publish_report()
         report.store_to_file(filepath)
 
     def set_comment(self, comment: str):
-        # Ignore change of comment when publishing started
-        if self._publish_has_started:
-            return
-        self._publish_context.data["comment"] = comment
-        self._publish_comment_is_set = True
+        self._logic.set_comment(comment)
 
     def run_action(self, plugin_id: str, action_id: str):
-        # TODO handle result in UI
-        plugin = self._publish_plugins_proxy.get_plugin(plugin_id)
-        action = self._publish_plugins_proxy.get_action(plugin_id, action_id)
-
-        result = pyblish.plugin.process(
-            plugin, self._publish_context, None, action.id
+        result: PublishActionResult = self._logic.run_action(
+            plugin_id, action_id
         )
-        exception = result.get("error")
-        if exception:
+        if result.success is False:
+            exception = result.exception
+            action = result.action
             self._emit_event(
                 "publish.action.failed",
                 {
@@ -827,105 +193,10 @@ class PublishModel:
                 }
             )
 
-        self._publish_report.add_action_result(action, result)
-
         self._controller.emit_card_message("Action finished.")
 
     def _emit_event(self, topic: str, data: dict[str, Any] | None = None):
         self._controller.emit_event(topic, data, PUBLISH_EVENT_SOURCE)
-
-    def _set_finished(self, value: bool):
-        if self._publish_has_finished != value:
-            self._publish_has_finished = value
-            self._emit_event(
-                "publish.finished.changed",
-                {"value": value}
-            )
-
-    def _set_is_running(self, value: bool):
-        if self._publish_is_running != value:
-            self._publish_is_running = value
-            self._emit_event(
-                "publish.is_running.changed",
-                {"value": value}
-            )
-
-    def _set_has_validated(self, value: bool):
-        if self._publish_has_validated != value:
-            self._publish_has_validated = value
-            self._emit_event(
-                "publish.has_validated.changed",
-                {"value": value}
-            )
-
-    def _set_is_crashed(self, value: bool):
-        if self._publish_has_crashed != value:
-            self._publish_has_crashed = value
-            self._emit_event(
-                "publish.has_crashed.changed",
-                {"value": value}
-            )
-
-    def _set_has_validation_errors(self, value: bool):
-        if self._publish_has_validation_errors != value:
-            self._publish_has_validation_errors = value
-            self._emit_event(
-                "publish.has_validation_errors.changed",
-                {"value": value}
-            )
-
-    def _set_max_progress(self, value: int):
-        if self._publish_max_progress != value:
-            self._publish_max_progress = value
-            self._emit_event(
-                "publish.max_progress.changed",
-                {"value": value}
-            )
-
-    def _set_progress(self, value: int):
-        if self._publish_progress != value:
-            self._publish_progress = value
-            self._emit_event(
-                "publish.progress.changed",
-                {"value": value}
-            )
-
-    def _set_publish_error_info(self, value: PublishErrorInfo | None):
-        if self._publish_error_info != value:
-            self._publish_error_info = value
-            self._emit_event(
-                "publish.publish_error.changed",
-                {"value": value}
-            )
-
-    def _default_iterator(self):
-        """Iterator used on initialization.
-
-        Should be replaced by real iterator when 'reset' is called.
-
-        Returns:
-            collections.abc.Generator[partial]: Generator with partial
-                functions that should be called in main thread.
-
-        """
-        while True:
-            yield partial(self.stop_publish)
-
-    def _start_publish(self):
-        """Start or continue in publishing."""
-        if self._publish_is_running:
-            return
-
-        self._set_is_running(True)
-        self._publish_has_started = True
-
-        self._emit_event("publish.process.started")
-
-    def _stop_publish(self):
-        """Stop or pause publishing."""
-        self._set_is_running(False)
-
-        self._emit_event("publish.process.stopped")
 
     def _publish_iterator(self) -> Iterable[partial]:
         """Main logic center of publishing.
@@ -1026,112 +297,3 @@ class PublishModel:
         self._set_finished(True)
         self._set_progress(self._publish_max_progress)
         yield partial(self.stop_publish)
-
-    @contextmanager
-    def _log_manager(self, plugin: pyblish.api.Plugin):
-        root = logging.getLogger()
-        if not self._log_to_console:
-            plugin.log.propagate = False
-            plugin.log.addHandler(self._log_handler)
-            root.addHandler(self._log_handler)
-
-        try:
-            if self._log_to_console:
-                yield None
-            else:
-                yield self._log_handler
-
-        finally:
-            if not self._log_to_console:
-                plugin.log.propagate = True
-                plugin.log.removeHandler(self._log_handler)
-                root.removeHandler(self._log_handler)
-            self._log_handler.clear_records()
-
-    def _process_and_continue(
-        self,
-        plugin: pyblish.api.Plugin,
-        instance: pyblish.api.Instance
-    ):
-        with self._log_manager(plugin) as log_handler:
-            result = pyblish.plugin.process(
-                plugin, self._publish_context, instance
-            )
-            if log_handler is not None:
-                records = log_handler.get_records()
-                exception = result.get("error")
-                if exception is not None and records:
-                    last_record = records[-1]
-                    if (
-                        last_record.name == "pyblish.plugin"
-                        and last_record.levelno == logging.ERROR
-                    ):
-                        # Remove last record made by pyblish
-                        # - `log.exception(formatted_traceback)`
-                        records.pop(-1)
-                result["records"] = records
-
-        exception = result.get("error")
-        if exception:
-            if (
-                isinstance(exception, PublishValidationError)
-                and not self._publish_has_validated
-            ):
-                result["is_validation_error"] = True
-                self._add_validation_error(result)
-
-            else:
-                if isinstance(exception, PublishError):
-                    if not exception.title:
-                        exception.title = plugin.label or plugin.__name__
-                    self._add_publish_error_to_report(result)
-
-                error_info = PublishErrorInfo.from_exception(exception)
-                self._set_publish_error_info(error_info)
-                self._set_is_crashed(True)
-
-                result["is_validation_error"] = False
-
-        self._publish_report.add_result(plugin.id, result)
-
-    def _add_validation_error(self, result: dict[str, Any]):
-        self._set_has_validation_errors(True)
-        self._add_publish_error_to_report(result)
-
-    def _add_publish_error_to_report(self, result: dict[str, Any]):
-        self._publish_errors.add_error(
-            result["plugin"],
-            result["error"],
-            result["instance"]
-        )
-
-    def _is_publish_plugin_active(self, plugin: pyblish.api.Plugin) -> bool:
-        """Decide if publish plugin is active.
-
-        This is hack because 'active' is mis-used in mixin
-        'OptionalPyblishPluginMixin' where 'active' is used for default value
-        of optional plugins. Because of that is 'active' state of plugin
-        which inherit from 'OptionalPyblishPluginMixin' ignored. That affects
-        headless publishing inside host, potentially remote publishing.
-
-        We have to change that to match pyblish base, but we can do that
-        only when all hosts use Publisher because the change requires
-        change of settings schemas.
-
-        Args:
-            plugin (pyblish.Plugin): Plugin which should be checked if is
-                active.
-
-        Returns:
-            bool: Is plugin active.
-        """
-
-        if plugin.active:
-            return True
-
-        if not plugin.optional:
-            return False
-
-        if OptionalPyblishPluginMixin in inspect.getmro(plugin):
-            return True
-        return False
