+++ client/ayon_core/tools/publisher/window.py
@@ -313,10 +313,10 @@ class PublisherWindow(QtWidgets.QDialog):
             "publish.process.started", self._on_publish_start
         )
         controller.register_event_callback(
-            "publish.has_validated.changed", self._on_publish_validated_change
+            "publish.has_validated", self._on_publish_validated
         )
         controller.register_event_callback(
-            "publish.finished.changed", self._on_publish_finished_change
+            "publish.finished", self._on_publish_finished
         )
         controller.register_event_callback(
             "publish.process.stopped", self._on_publish_stop
@@ -974,14 +974,12 @@ class PublisherWindow(QtWidgets.QDialog):
         if self._is_on_create_tab():
             self._go_to_publish_tab()
 
-    def _on_publish_validated_change(self, event):
-        if event["value"]:
-            self._validate_btn.setEnabled(False)
+    def _on_publish_validated(self):
+        self._validate_btn.setEnabled(False)
 
-    def _on_publish_finished_change(self, event):
-        if event["value"]:
-            # Successful publish, remove comment from UI
-            self._comment_input.setText("")
+    def _on_publish_finished(self, event):
+        # Successful publish, remove comment from UI
+        self._comment_input.setText("")
 
     def _on_publish_stop(self):
         self._set_publish_overlay_visibility(False)
